move concurrency test to PidGeneratorAppSpec
